no termina de funcionar la validacion

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -46,7 +46,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setFloat(idx, comercial.getComision());
+			ps.setBigDecimal(idx, comercial.getComision());
 			return ps;
 		},keyHolder);
 
@@ -65,7 +65,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                 							  rs.getString("nombre"), 
                 							  rs.getString("apellido1"),
                 							  rs.getString("apellido2"), 
-                							  rs.getFloat("comisión"))
+                							  rs.getBigDecimal("comisión"))
                 						 	
         );
 		
@@ -84,7 +84,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 								rs.getString("nombre"),
 								rs.getString("apellido1"),
 								rs.getString("apellido2"),
-								rs.getFloat("comisión"))
+								rs.getBigDecimal("comisión"))
 								, id
 						);
