+++ src/main/java/org/iesvdm/dto/ComercialDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -14,7 +16,7 @@ public class ComercialDTO {
     private String nombre;
     private String apellido1;
     private String apellido2;
-    private float comision;
+    private BigDecimal comision;
 
     // Atributos auxiliares de Pedido
     private double importeTotalPedidos;
